+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 * Add `EveOnline::ESI::UniversePlanet`
 * Add `EveOnline::ESI::Models::AsteroidBelt`
 * Add `EveOnline::ESI::UniverseAsteroidBelt`
+* Add `EveOnline::ESI::UniverseMoon`
+* Add `EveOnline::ESI::Models::Moon`
 
 ## v0.19.0